Created "supla-common" folder and moved common files there
