+++ libraries/SuplaDevice/supla-common.c
@@ -1,0 +1,8 @@
+/*
+ *  Workaround for lack of nice build system in Arduino
+*/
+
+#include "supla-common/proto.c"
+#include "supla-common/lck.c"
+#include "supla-common/srpc.c"
+
