+++ libraries/SuplaDevice/supla-common/IEEE754tools.h
