+++ libraries/SuplaDevice/supla-common/log.h
