+++ libraries/SuplaDevice/supla-common/proto.h
