+++ libraries/SuplaDevice/supla-common2.cpp
@@ -1,0 +1,7 @@
+/*
+ *  Workaround for lack of nice build system in Arduino
+*/
+
+
+#include "supla-common/log.cpp"
+
